audio-streamer/requirements.txt
pyaudio==0.2.14
google-cloud-speech==2.27.0
google-cloud-translate==3.16.0
google-cloud-pubsub==2.*

setuptools
